chore(volume): remove volume operations
The REST API and agents should only support operations on primitive
types. A volume is an abstract concept and was included as a convenience
to the consumer of the API. However, the assumption is that the consumer
has significant domain knowledge (i.e. an operator) and therefore should
not require this.