feat(persist child states): use etcd as the store
When a child changes state, persist the new state to the store. This
ensures that when a nexus is recreated it can check the state of the
children before deciding whether or not to use them.

Currently only healthy children are added to the nexus on creation. The
expectation is that the control plane will eventually replace the
unhealthy children in order to satisfy the desired replica count.