chore(tidy up): persisting child states
Provide a function which can be used to persist child states to the
store. For now this function continually tries to store the child states
on failure.

Add additional persistent store errors which are used when there is a
failure to wait on a channel for a completed store operation.