test(persistence): add test for persistence
The test checks that the child state is stored to etcd when it is
retired. It is marked as "ignored" for now because it is waiting on
other changes before it will succeed.

Added an additional proto file. This is a draft version of the structure
that will be store in etcd. It is defined in a proto file so that both
Mayastor and MOAC can use it.