feat(persistent struct): define persistent info
The data structure stored in the persistent store has been defined in a
new persistence.proto file. The reason for defining it here is because
this structure will be common to MOAC and Mayastor.

A clean shutdown bit is also included in this structure. When a nexus is
destroyed, the clean shutdown variable is set to true. In the event of a
crash, this variable will be false.

On nexus creation, the data structure in the persistent store is
overwritten with the latest information.