feat(etcd reconnect): retry connections
When etcd is down, the etcd-client library will hold on to any
outstanding requests and waits for a TCP timeout of 15 minutes before
re-establishing a connection. Rather than waiting for this, add a 10
second timeout to each store operation and try to create a new
connection to etcd if it times out.